refactor: cartController와 cartService 코드 최적화 (테스트 다깨짐)

+++ src/main/java/cart/cart/presentation/dto/DeliveryDiscountResponse.java
@@ -1,13 +1,13 @@
 package cart.cart.presentation.dto;
 
 public class DeliveryDiscountResponse {
-    private long originalPrice;
-    private long discountPrice;
+    private int originalPrice;
+    private int discountPrice;
 
     public DeliveryDiscountResponse() {
     }
 
-    public DeliveryDiscountResponse(long originalPrice, long discountPrice) {
+    public DeliveryDiscountResponse(int originalPrice, int discountPrice) {
         this.originalPrice = originalPrice;
         this.discountPrice = discountPrice;
     }
